+++ CHANGELOG.md
@@ -8,6 +8,7 @@ All notable changes to the Aptos TypeScript SDK will be captured in this file. T
 - Rename custom request header to `aptos-typescript-sdk`
 - [`Breaking`] Rename `token` to `digitalAsset` and add digital asset built in transaction generation functions
 - [`Breaking`] change transaction submission builder flow namespace to be under a `transaction` namespace
+- [`Breaking`] Rename `SingleSignerTransaction` type to `SimpleTransaction`
 
 ## 0.0.8 (2023-11-29)